refactor: break server.js into modular routes and lib modules (#7)
BREAKING CHANGE: server.js has been completely restructured.

- Split monolithic server.js (1600+ lines) into modular architecture:
  - src/lib/state.js — shared state (config, alertConfig, clients, etc.)
  - src/lib/gateway.js — gateway process detection, kill, restart
  - src/lib/alerts.js — webhook/Telegram alert dispatching
  - src/lib/streaming.js — external log sink streaming
  - src/lib/sequence-helpers.js — suspicious sequence detection
  - src/routes/activity.js — /api/activity, /api/stats
  - src/routes/alerts.js — /api/alerts/config GET/POST
  - src/routes/config.js — /api/config GET/POST
  - src/routes/dump.js — /api/dump session export
  - src/routes/exports.js — /api/export/json, /api/export/csv
  - src/routes/gateway.js — /api/gateway/status, kill, restart
  - src/routes/sessions.js — /api/sessions
  - src/routes/streaming.js — /api/streaming config/status
  - src/routes/version.js — /api/version (update checker)
- New server.js is ~100 lines (imports + Express mount + chokidar watcher)
- Old server.js preserved as server-old.js for reference
- Expanded test suite: 17 tests (was 12)
  - Added: dashboard HTML, config GET/POST, dump preview, WebSocket
  - Removed unsafe gateway kill/restart tests (they kill real processes)
- Tests now use CLAWGUARD_TEST_PORT env var (default 3847, CI uses 13847)
- CI workflow now starts server and runs npm test
- Updated syntax check script to cover all new modules
- Updated .npmignore to exclude server-old.js and REFACTOR.md
